requirements.txt
pyTelegramBotAPI
python-a2s
python-dotenv